+++ components/navigation/components/navigation-item.tsx
@@ -19,6 +19,7 @@ export const NavigationItem = ({
     <Link
       href={href}
       className={cn(
+        "hover:text-xl",
         pathname === href
           ? "text-blogThickRed"
           : "hover:text-blogThickRed",
